requirements.txt
streamlit
tensorflow-cpu==2.12.0
gdown
numpy
Pillow
